safari-600.7-branch/LayoutTests/platform/ios-sim/fast/text/data-detectors/phone-number-parse-setInnerHTML-expected.txt

Test that there is no phone number parsing via innerHTML

no phone number detection:
| "
    "
| <iframe>
|   id="iframe"
|   src="resources/phone-dynamic.html"
| "


"

FRAME 0:
| <head>
|   "
"
| "
"
| <body>
|   "
  "
|   <div>
|     id="container"
|     "Call me (308) 555 5555 please"
|   "


"
